as we no longer expect `off_t` to be stored, the type should better be `uintptr_t`

+++ include/h2o/socket.h
@@ -168,7 +168,7 @@ struct st_h2o_sendvec_t {
      */
     union {
         char *raw;
-        uint64_t cb_arg;
+        uintptr_t cb_arg;
     };
 };
